feat: Added ProfileState, added resetError method to userState
