+++ src/store/exports.js
@@ -3,5 +3,6 @@ export { default as catalog } from './catalog';
 export { default as modals } from './modals';
 export { default as article } from './article';
 export { default as locale } from './locale';
-export { default as user } from './user'
-export { default as category } from './category'
+export { default as user } from './user';
+export { default as profile } from './profile';
+export { default as category } from './category';
