+++ src/store/user/index.js
@@ -9,7 +9,7 @@ class UserState extends StoreModule {
       data: {},
       waiting: false, // признак ожидания загрузки
       error: null,
-      isAuth: false
+      isAuth: JSON.parse(localStorage.getItem('isAuth'))
     };
   }
 
@@ -41,6 +41,7 @@ class UserState extends StoreModule {
       }
 
       localStorage.setItem('userToken', json.result.token)
+      localStorage.setItem('isAuth', true);
 
       this.setState({
         ...this.getState(),
@@ -85,6 +86,7 @@ class UserState extends StoreModule {
       }
 
       localStorage.removeItem('userToken')
+      localStorage.removeItem('isAuth');
       
       this.setState({
         ...this.getState(),
@@ -102,48 +104,11 @@ class UserState extends StoreModule {
     }  
   }
 
-  /**
-   * Загрузка информации о профиле
-   * @return {Promise<void>}
-   */
-  async loadUserInfo() {
+  resetError() {
     this.setState({
       ...this.getState(),
-      error: null,
-      waiting: true,
-    })  
-    try {
-      const token = localStorage.getItem('userToken')
-      if (!token) {
-        throw new Error('Токен не найден')
-      }
-      const response = await fetch('/api/v1/users/self?fields=email,username,profile(name, phone)', {
-        method: 'GET',
-        headers: {
-          'Content-type': 'application/json',
-          'X-Token': token
-        }
-      })
-      const json = await response.json()
-      
-      if(!response.ok) {
-        throw new Error(json.error.data.issues[0].message)
-      }
-
-      this.setState({
-        ...this.getState(),
-        data: json.result,
-        waiting: false,
-        isAuth: true,
-      }, 'Информация о пользователе загружена из АПИ')
-    } catch (error) {
-      // Удаление испорченного токена
-      localStorage.removeItem('userToken')
-      this.setState({
-        ...this.getState(),
-        waiting: false,
-      })
-    }
+      error: null
+    })
   }
 }
